exception handling and validation has added

+++ src/main/java/com/example/demo/Service/BookService.java
@@ -1,5 +1,6 @@
 package com.example.demo.Service;
 
+import com.example.demo.ExceptionHandling.ResourceNotFoundException;
 import com.example.demo.Model.Book;
 import com.example.demo.Repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +19,8 @@ public class BookService {
     }
 
     public Book getBookById(int id) {
-        return bookRepository.findById(id).orElse(null);
+        return bookRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Book not found with id:" + id));
     }
 
     public Book addBook(Book book) {
@@ -26,6 +28,21 @@ public class BookService {
     }
 
     public void deleteBookById(int id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Book Not found with this id:" + id);
+        }
         bookRepository.deleteById(id);
+    }
+
+    public Book updateBookById(int id, Book book) {
+        Book existingBook = bookRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Book not found with id: " + id));
+
+        existingBook.setTitle(book.getTitle());
+        existingBook.setAuthor(book.getAuthor());
+        existingBook.setPrice(book.getPrice());
+
+        return bookRepository.save(existingBook);
     }
 }
